Update async test for client v2
